fixing bugs and resolving issues

+++ src/main/java/com/crud/library/controller/CopiesController.java
@@ -51,7 +51,7 @@ public class CopiesController {
     }
 
     @DeleteMapping(value = "deleteCopy")
-    public void deleteCopy(@RequestParam int copyId) {
+    public void deleteCopy(@RequestParam Long copyId) {
         copiesDbService.deleteCopy(copyId);
     }
